requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
file_read_backwards==3.1.0
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lgpio==0.2.2.0
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
smbus2==0.5.0
urllib3==2.3.0
Werkzeug==3.1.3
